Refactor: Restrctured the upload CV component
