+++ client/src/Components/sectionTwo/SectionTwo.js
@@ -1,24 +1,14 @@
-import React, { useState} from "react";
+import React from "react";
 import { useGlobalContext } from "../../context/context";
+import Upload from "../upload/Upload";
 import Uploading from "../uploading/Uploading";
-import first from "./assets/first.svg";
+
 
 function SectionTwo() {
-	const [fileSize, setFileSize] = useState();
-	const { setFile, setFileName } = useGlobalContext();
+	
+	const { fileSize } = useGlobalContext();
 
-	const changeHandler = (e) => {
-		setFile(e.target.files[0]);
-		setFileSize(e.target.files[0].size);
-		setFileName(e.target.files[0].name);
-	};
-	const fileValidation = () => {
-		return (
-			<h2 className="text-errorMain font-semibold text-[13px]">
-				The file too large. Please Upload a file with maximum of 5mb
-			</h2>
-		);
-	};
+	
 	return (
 		<div
 			id="upload_section"
@@ -40,45 +30,11 @@ function SectionTwo() {
 					}  border-dashed rounded-lg `}
 				>
 					<div className="uploadContainer relative flex flex-col items-center justify-center lg:px-[2vh] lg:py-[15vh] py-[15vh]">
-						{fileSize > 0 && fileSize < 5000000 ? null : (
-							<img
-								src={first}
-								className="w-[67px] lg:left-[37%]  h-[67px]"
-								alt=""
-							/>
-						)}
-						{fileSize > 0 && fileSize < 5000000 ? null : (
-							<input
-								data-testid="element"
-								style={{ opacity: "0" }}
-								type="file"
-								name="file"
-								accept="application/pdf"
-								onChange={changeHandler}
-								className="upload_file absolute cursor-pointer left-0 bottom-[10%] w-[100%] h-[80%] "
-								id="upload_button"
-							/>
-						)}
-
-						{fileSize > 0 && fileSize < 5000000 ? (
-							<Uploading />
-						) : (
-							<label htmlFor="upload_button" className="flex flex-col items-center justify-center">
-								<img src="" alt="" />
-
-								<h3 className=" text-primaryMain lg:text-[24px] text-center mt-4 text-[20px] font-semibold">
-									Drag & Drop to Upload{" "}
-								</h3>
-
-								<p className="text-grey300 text-[16px] text-center">
-									File Supported: PDF
-								</p>
-								{fileSize > 5000000 ? fileValidation() : null}
-								<button className="text-primaryMain text-[16px]  mt-3 text-center font-bold ">
-									Or browse
-								</button>
-							</label>
-						)}
+						{
+							fileSize < 5000000 && fileSize > 0  ? <Uploading /> :
+							<Upload/> 
+						}
+					
 					</div>
 				</div>
 			</main>
